Write func to send program from server

+++ pi-side/server.h
@@ -11,8 +11,8 @@ uint32_t all_pis();
 // PIs currently broadcasting their availability.
 uint32_t free_pis(uint32_t pi_id);
 
-// Send a program from the sesrver to client pi_id
-void send_pi_prog(uint32_t pi_id, uint32_t nbytes, void *code);
+// Send a program from the server to client pi_id
+void send_pi_prog(uint8_t pi_id, uint32_t nbytes, void *code);
 
 // ESP maintains return values from client broadcast_avail
 // When the server discovers that a new pi is free, calls
